Add Van-Leer limited heated channel test
The result from this test shows a beautiful flat profile for the
momentum. It has to be run transiently because I think VanLeer steady
struggles with constant initial data (I believe we get a/some singular
values in our matrix). That requirement led to creation of the
`FVMatPropTimeKernel` in this commit. Note that along with the beautiful
flat momentum profile, the nonlinear variable set (pressure, temperature
and superficial velocity) along with rho are perfectly monotone and
without any boundary artifacts. I have to say it's beautiful :-)