allow app-custom src dependencies during compilation
This is useful for apps that may have moose-external dependencies that
they need to compile/build.  Sometimes a moose app file (e.g.
FooMaterial.C) can depend on a file generated by the build of an
external app dependency.  Without this addition - there is not way for
the app developer to tell us about this dependency - which prevents
parallel builds from working properly.

Fixes #17800.