Add a "backup" FunctionBase to CachedData
This lets us use our own pre-existing solution as backup, which in turn
lets us do proper projections onto target elements which are
incompletely overlapped by the source domain.

Of course, the proper thing to do on target elements which are not *at
all* overlapped by the source domain is to skip the projection entirely;
that will have to wait for future optimization work.