Add porosity jumps plus friction min-mod KT test
I learned quite a bit while compising this test. It appears that in the
absence of friction if you go from having a continuously varying
porosity to a discontinuous jump then you have an irreversible pressure
drop. So for instance if you step down in porosity and then step back up
to the original value, your pressure (at least in all my simulation
attempts) will finish lower than it started and correspondingly your
superficial velocity will be higher. However, if you do the same process
continuously your pressure returns to its original value. I don't know
what to make of it. I doubt that the discontinuous result is right.
However, I can say that with any significant frictional drag force, the
effects of the porosity jumps on the pressure (and density) are
miniscule compared to the effects of drag. So for something like HTR-PM
the interesting behavior with the pressure field due to porosity jumps
may not matter