Update contact mechanics action to include variationally consistent contact constraints with weighted quantities using dual bases.
The action allows for having two cases: "weighted", which uses weighted gaps and tangential velocities and dual bases; and
"legacy", which uses a combination of nodal and mortar constraint enforcement via 'NormalNodalLM' and 'TangentialMortarLM' objects,
mixing the interpolation orders of standard bases.

The performance of both approaches is very similar, but a small deterioration of the Jacobian when using the default 'weighted' approach
is possible. However, to get the full benefits of the mortar approach, selecting the default 'weighted' approach is strongly recommended.

Refs #16961, #17495