Address Alex's comment.
Decreasing `l_tol` does not guarantee better convergence. Especially for this test, where the true residual goes up during
the first nonliner iteration. Given that the first nonliner iteration is responsible for the sporadic cuts in the time
step, I am reverting here the modification of this value. I am keeping the increase in nonlinear iterations, as in those
instances where the test fails, it seems right on track to convergence in a few more nonlinear iterations.