Fix CIVET result links for master branch documentation
The fix in #17178 was operting for the PR/devel levels but not linking to the correct pages on
master. This latest change will result in the CIVET badges only being available on the master
documentation. This is the correct behavior, because that is the documentation intented for public
consumption.

(refs #17177)