Generalized flow channel alignment check to allow gap
An error test was changed because the generalization allows the
particular case of transversely translated flow channel. To re-trigger
the error, the flow channel was shifted axially, and then it was
also needed to refine the mesh since the 2-element discretization
was not fine enough to alter the mapping between flow channel
elements and their neighbors.

Closes #19878