Avoid regold in geomsearch grid sequencing test
The issue is somewhat complex. receiver is an AuxVariable and it's
filled with a boundary restricted aux kernel. Previously aux kernels
executing on timestep end would be executed for each stage of the grid
sequence but now they only execute for the very last grid. Consequently
for the first time step only the nodes on the boundary will have
non-zero values for the aux variable after Yaqi's change. However,
if we change the input to also execute this boundary restricted aux
kernel on nonlinear then we restore the previous behavior with respect
to exodus output:

1. auxkernel is evaluated on the coarse grid
2. refinement occurs and the boundary nodal solution on the coarsest
   grid is projected such that the first layer of interior nodes
   also has non-zero values for the aux variable
3. auxkernel is computed on the finer grid and the boundary nodal
   values are updated
4. solution is output and both the boundary nodes and the first layer
   of interior nodes have non-zero aux variable values