Remove block insertion check
This check was immediately triggered because we add three material
objects per material in the input file, e.g. volumetric, face, and
neighbor face materials. Without the check we are vulnerable to user
error such as defining two materials in the input file that define the
same functor property on a given block. Maybe the best remedy in the
future is to have distinct `FunctorMaterials` that don't get added in
triplicate because there is no functional reason for that in a system
where we are evaluating material properties on the fly as opposed to
having references to triplicate storage