Put --no-as-needed in linker flags
This fixes #16092 for me on Ubuntu 21.04, where presumably the default
linker behavior is --as-needed?  As explained in that ticket: when we
only interact with a library by invoking the constructors of its static
initialized objects, --as-needed breaks our interaction by failing to
determine that the library has to be loaded at all.

This flag isn't supported on OSX (and perhaps nothing equivalent is,
according to a brief search), so we test to see whether trying to use it
throws an error and we only enable it if not.