Working wcnsfv natural convection example
It was actually working before but when there are large variations
in density then the velocity distribution is asymmetric which is unlike
the Boussinesq benchmarks that I'm used to looking at (which kind of
makes sense... Boussinesq is only supposed to be applied when density
variations relatively small). So instead of generating Rayleigh numbers
by varying delta T (and consequently delta rho), I generate Rayleigh
numbers based on changing L (e.g. xmax and ymax)