Restore aggregate interpolation in FVMatAdvection
This removes the previous exodiff. This also adds a new overload
of operator() that combines an element, face, and subdomain ID. For
variables we check whether we have dofs on the element, and if not
we create a ghost value using the element across (using the face)
and any Dirichlet information. We use the subdomain ID for ghosted
material properties