Address Daniel's comments (#16016)
Avoid using Euler angle sequences and perform the transformation of the 6-by-6
Hill's tensor directly from the components of the rotation matrix resulting from
large deformation rotation and a possible initial orientation.

Now, we avoid extracting Hill constants when we transform the Hill's tensor. The
default value for the HillConstant material is to compute the transformation,
and the default for the creep model is to use such a transformation. To support
simple cases where changes of orientation aren't needed, the use can select
use_transformation = false in the creep model and use_large_rotation = false
in the HillConstant material.