Avoided accidental use of nullptr in ExplicitSSPRungeKutta::computeADTimeDerivatives
A quiet NaN is used in cases where a nullptr is attempted to be used.
If the result value is actually used in a residual or Jacobian, then
this case will be identified, and a different solution will be
implemented; however, the current approach was found to be sufficient
for the input file attached to the issue, thus indicating that the
offending call produced a value that was never actually used, as
expected.

Closes #18643