Always compute single point map when computing Jacobian
As shown in libmesh/libmesh#2997, when the threshold for
determining non-affine is very low, e.g. if we make our affine
determination to a much tighter tolerance, then it's possible
to see discontinuities in a contrived test like the one changed here.
As described in the referenced libmesh pull request, if elements
are affine, then the user Jacobian evaluation routine will treat them
as such, but a slight perturbation to the displacement field, as is done
in finite difference testing of the Jacobian, will make the element
non-affine and consequently proceed down a different code path in
function evaluation. The solution is to make sure that when performing
Jacobian evaluations and our affine map determination is a function of
nonlinear degrees of freedom (e.g. the displacement variables) that we
always compute the single point map