Remove overzealous assert in ADComputeFiniteStrain
See current failure on idaholab/blackbear#250. Test passes in opt
(with no asserts) and fails in debugging modes because of the assert.
It's much better not to assert and instead to insert a NaN into the
residual vector. PETSc will quite naturally cut the lambda in the
line search when it gets a NaN in its fnorm testing which is exactly
what we want here. We want to reduce the Newton step such that we
stay within physical bounds. Right now we don't do that and instead
we just abort. There's absolutely no need for that