Always clone rms
Also continue the very slow march to make relationship manager code more
readable which hopefully also means less liable to bug-ridden. This
change came about because on Jacob's 3D mortar PR he has a test that
collects two meshes into a single one using the
`MeshCollectionGenerator`. We were failing an assertion when running
with distributed mesh because a relationship manager/ghosting functor
was using the incorrect `MeshBase` object. To avoid this we should
always clone and make sure we're initializing with the correct mesh

Refs #18203