Add tests for arbitrary eigenstrain gradient in interaction integral ref #18804
Add a test using a testing material that computes an eigenstrain
gradient as a RankThreeTensor and gives a very similar result to
an existing test that does the same calculation but through the thermal
gradient. The results are slightly different because of differences in
interpolation of the driving eigenstrains.

Also add error condition tests for this new capability.