sets of pointers hurt reproducibility
Depending on the memory layout a given process may store objects in
a different order from simulation run to run. By comparing against
object id, we can be sure that the simulation will behave identically
(with respect to how this particular data structure is sorted and
in any code whose behavior is determined by the sorting of said structure)
from run to run.

This commit also removes a bunch of non-gold exodus output files

Refs #19092