Try to limit total number of relationship managers added
If an RM asks to be run on the displaced mesh, then we should respect
that and only add it for the displaced mesh. Because in theory that RM
may only function properly if it's really acting with the displaced mesh
(or the converse may be true: a RM may only function properly if it's
acting on the reference mesh). Our proxy geometric relationship managers
should handle ghosting for whatever mesh the RM *doesn't* get added to.
So a summary of what this commit does:

- The same way we only add algebraic ghosting functors to the
  requested undisplaced/displaced dof-map, we now only add geometric ghosting
  functors the requested undisplaced/displaced mesh
- We allow geoemetric proxy relationship managers (which we did not add
  before) to ensure that ghosting is consistent between reference and
  displaced meshes, whereas before we just relied on the idea that cloning
  geometric ghosting functors and hoping they gave the same ghosting would
  work out for us. Thus far in our test suite it had, but who knows what
  would have happened in the future
- With the new existence of geometric proxy relationship managers, we
  needed to make some changes in the `Adaptivity` class in order to
  ensure that the meshes stay roughly in sync such that the
  `ProxyRelationshipManager` can function correctly (e.g. it relies on
  unique id queries so elements cannot be deleted or renumbered before
  both meshes have been adapted)