New class works fine with nonsparse local AD
With global AD indexing we always force a full preconditioning matrix
because we have no way to know what global indices correspond to what
variables without building a map. However, with local numbering we do
know that information so we don't force a full preconditioning matrix.
Consequently convergence will be sensitive to what a user specifies in
their input. In this case, no preconditioning matrix is specified so for
the nonsparse case you get a block diagonal. With different matrices,
the global and local setups converge differently and it just so happened
that the global converged to a lower norm than the local case. By
scaling the temperature and LM we force both variables to finish the
simulation converged, and this gives us consistent results across AD
types