Revert "Try to limit total number of relationship managers added"
See https://github.com/idaholab/moose/pull/19349#issuecomment-965805132
for discussion and reasoning behind this. I'll reproduce the comment
below:

Ok Fande sets `skip_partitioning(true)` in his distributed mesh generator. But
when we recover `_skip_all_partitioning` is set to `false`. We could debate
whether we should write out that data member to checkpoint so that the flag is
the same upon recover as it is before recover, but that's not the main point
here ...

When we have both distributed and displaced mesh, I think that geometric proxy
relationship managers simply won't work when repartitioning is allowed. When the
reference mesh is repartitioned it will redistribute its elements. Then when we
call repartition on the displaced mesh, we will be querying its ghosting
functors, one of which will be the geometric proxy relationship manager. This
will look over at the reference mesh and run its ghosting functors with
`reference_mesh.active_local_elements_begin()` and
`reference_mesh.active_local_elements_end()` as calling args and return a set of
elements. What is happening in the failing test is that the returned set has
elements that do not yet exist on the displaced mesh (as ascertained through
checking unique id) because we *have not yet* redistributed the displaced mesh;
we are in fact querying the ghosting functors to help us determine what elements
to send/receive in our redistribution. And because this reference element does
not yet have its displaced mesh element version on the current process, we fail
the `unique_id_to_elem_map` find assertion.

I think my conclusion here is that proxy relationship managers are not the
solution for maintaining synchronicity between meshes. I'm going to drop that
code from this PR. This kind of stuff really makes me want to get
https://github.com/libMesh/libmesh/issues/2661