Added ability to use closures created in input file to flow channels
The ability to use enumerated closures options has been preserved.
If the provided closures parameter corresponds to closures created
via the input file, then those closures are used. Else the enumerated
options are tried; if it corresponds to an enumerated option, that is
used, and a deprecated message is given. If the closures parameter
does not correspond to either method, then an error occurs.

Closes #19682