Elem::centroid() -> Elem::vertex_average()
libMesh deprecated Elem::centroid() due to its misleading name - that
method was actually returning a vertex average for most elements, even
when that average was not the true centroid.

For this application it looks like vertex_average() (which is faster
than a true_centroid() calculation in general) should work correctly;
let's request it explicitly.

Refs #13921