Swap-back sentinels for boundary aux kernels
If an exception gets thrown out of reinitMaterialsXXX or callees of
reinitMaterialsXXX then none of the remaining code in
ComputeElemAuxBcsThread::operator() will get executed, e.g. in the old way,
_problem.swapBackMaterialsXXX would never get called. However, the C++ language
is nice enough that, even though it won't execute any following lines of code,
it will destroy any scope-local variables, like these SwapBackSentinels. And in
the destructors of SwapBackSentinel we call problem.swapBackMaterialsXXX.

This swap-back is necessary for stateful material properties ... essentially if
you don't swap-back after swapping then you'll be likely to be operating on
material properties in the future that aren't "stateful"; e.g. you'll be working
with some null, uninitialized dummy quantity. This was the case here ... the
sifgrs property never got swapped back so you wound up operating on some dummy
quantity that had never been initialized/properly-sized

Refs #000 for now