Don't prepare_for_use in mortar segment mesh
We could potentially need remote mortar elements for our new
`MortarNodalAuxKernel`. This need would arise any time a node is on a
processor boundary and consequently the aux kernel would call for
evaluations of non-local mortar segments associated with that node. And
if we are not deleting remote elements, partitioning, or finding
neighbors, then I don't really know why we should call `prepare_for_use`