Use an ordered set based on ID for secondary to msm container
I ran into occasional failures about reusing the base for matrix-free
computations on CIVET and in repeated running locally
(strangely only when running valgrind (which I should note never
errored, it just was the only thing that helped trigger the libMesh
failed assertion)). After some bisecting I tracked the change to moving
from a vector sequence based on the msm ids to an unordered set based
only the msm pointers. Because the msm mesh is rebuilt upon ever
residual evaluation the address space used by the msm elements is liable
to change and so the hashing and consequently the container iteration
order is liable to change. This is what led to the reuse base errors.
Because I always expect the number of mortar segments per secondary to
be relatively small I think it makes cpu sense to switch to an ordered
container and this has the happy side-effect of making the container
iteration order deterministic.

Perhaps here is a good place to point out that while the secondary to
msm container is an unordered_map, the secondary element pointers do not
change from residual to residual and hence the hashing and container
iteration order do not change, so that is why we have never seen an
issue related to that "unordered-ness" before

Refs #20017