Insert dependencies from discrete materials
Currently when we retrieve a discrete materials, we touch nothing about its dependencies. This leads to a weird situation where, when a discrete material is called by its hosting material, its dependencies may have not been evaluated. To avoid this situation, I propose to add the materials requested by the discrete material into the hosting material trying to retrieve the discrete material. I am not restricting the retrieval of discrete materials to the initial setup, so that we only need one more dependency resolution after the initial setup.

close #18726