Investigate use of thread-local static do_derivatives
The result, for the 2D turbulent WCNSFV channel-flow test with a couple
levels of mesh refinement, was zero speedup. Given the added code
complexity of having the static be thread-local, I would say zero
speed-up is not worth it. Maybe it's possible that a 3D case with more
sparsity pattern union computation (e.g. more columns per row) would
show a gain with thread-local, but until we have a demonstration of
speedup, this commit is not worth it.

Moreover, my profiling continues to show that the single heaviest
function is the PC application so that's the prize we should be focused
on