Really do compute the residual and Jacobian together
It turns out that the movement of residual and Jacobian object setting
from the constructor to init was very intentional. This is because
FEProblemSolve construction has not occurred at the time of nonlinear
system construction. With this new commit, when FEProblemSolve calls
NonlinearSystemBase::residAndJacobianTogether we toggle from using
tradtional residual and Jacobian to the joint object. This should be
compatible with PBP which were the tests that failed when I had moved
object setting to init.

This commit also introduces a new processResiduals API meant only for
processing residuals (and not their derivatives/Jacobian). I did this
because I was getting failures in NDEBUG modes about different residual
evaluations between the JFNK residual evaluation and the nonlinear residual
evaluation. This was due to the fact that we were caching per element
per residual object for the nonlinear residual and per element (all
residual objects) for JFNK residual. The new processResiduals API allows
residual objects to follow similar patterns for their residual and Jacobian
evaluation implementations.

When I added the new processResiduals, I didn't initially restrict it to
global AD indexing and I got an error on CIVET concerning the lack of
definition of `_scaling_vector` in the processResiduals implementation.
This led to me adding a `scaling_factor` argument to call the vector
of residuals APIs I have on Assembly, such that the methods can be used
equally well by both non-global and global AD. I'm not ready to give up
totally on non-global, nonsparse AD because Gavin has shown it's
demonstrably faster than global, sparse for our INSAD navier-stokes
implementation.

This regolds a THM test that was previously failing at time step 334. I
don't think that my changes in this commit made the solve any better; I
think this test was just sitting right on the edge