Fix MortarFrictionalPressureVectorAux for edge cases.
Particularly, when the lower-d mesh does not extend to the entire outer
topology, we may locate an internal element from a node on the secondary surface
that does not have a corresponding lower-d secondary element since it happens to be
outside the contact sideset. Since we were not checking for this, a bad access can
happen, and it does.

Now, we are sure we are obtaining one of the possible lower-d secondary elements.
Still, it doesn't matter which one we are getting since we loop over nodes, so
 any lower-d element does the trick.