Address Alex's review. (#19671)
Communicate wear depth among processors. We have to do this because
we are taking the wear depth computed from the mortar auxiliary kernel and then
integrate it over the mortar segments, which may reside in different processors.
This strategy may not be strictly necessary but it fits right in with the existing
design of mortar contact mechanics constraint classes.