Fixing up tests
Getting rid of assumptions in tests that rely on the file structure further up than the copy point. This
includes the relative location of the executable.

Also removing the empty executable option. While this isn't an unreasonable test, we don't have a good
capability today to run a test that relies on a relative path outside of the installed tree. We may re-add
this capability at a later date. By switching all the useability tests back to RunApp we can't control the
CLI arguments that are passed to the application, invalidating the need for that test.

refs #19022