Implement finite volume functor argument caching
Got to use structured bindings and `try_emplace` for the first time.
I'm a little concerned about the cost of binary trees but I feel like
hashing of the finite volume functor arguments could be even more expensive.
I think if this shows up as an important piece of a profile, then we
can investigate an unordered_map implementation

Closes #20470