Trim ADReal derivative size for Lagrange elements in mortar (#20488)
There is a risk in trimming derivatives since derivatives with small
absolute values can potentially be meaningful and necessary for PJFNK
and NEWTON convergence. However, in practical problems, we have seen that
ADReal can sometimes have derivatives on the order 1.0e-23 ~ 1.0e-27, which
are well below the machine epsilon and turn out to be many orders of magnitude
lower than other derivative sets. Since accounting for derivatives isn't free,
this commit adds a handle for removing derivatives that may not be numerically
relevant. An option to change the threshold value is provided.