Interior node AD derivative removal. (#20488)
When coupling AD variables into mortar constraints acting on Lagrange
elements, the interior nodes carry derivative information that may be
useless to the mortar contact problem. Since the number of derivatives
needed for a mortar problem can be large (e.g. 3 displacements + temperature
+ 1 LM for normal pressure + 2 LM for frictional pressure + 1 LM for heat flux),
a problem that is exarcebated by secondary-primary projections, the total number of
derivatives required can balloon (to the point of running out of stack).

Removing unnecessary interior node derivatives can cut in a third this
number. For example, for a HEX27 element, we'd only require derivative info
pertaining to 9 nodes, instead of 27.