Feel like I'm getting into hacking territory
It turns out that there are some baked in assumptions of where the App is located
relative to the testroot. There are some nasty hacks to locate it in a normal
installation location, but this doesn't work if you install elsewhere (like we have
to for the test). In order to get the TestHarness to behave, it's easiest to just
link the executable to the fake test hierarchy. This part isn't a hack ;)

Next Problem, the TestHarness currently checks for the working directory to exist
in the Runnable phase, which occurs asyncronously to the scheduled launch of the
test. This is a no go, since we are trying to copy the directory structure in the test
so it won't exist for latter tests. So... I moved the working directory check to
the place where we actually launch the test. This would be OK, but apparently there
are other assumptions baked into the Job class. I had to start/stop the timer to
avoid an index out of range.

Now the tests are running properly.

refs #19022