Add framework coverage of trim*Derivative
Some API changes as well:
- use std::vector<ADReal> instead of std::vector<ADReal *>
- template trimInteriorNodeDerivatives to allow for multiple types
  of containers

I want to allow for multiple types of containers because I don't
want users having to do dynamic memory allocation in methods like
`computeQpProperties`. I feel like that could be slow. So that means
allowing `std::array` ... which lead me to undo a change I forced
Antonio to make. Now we allow the variable containers to include
nullptrs for logical LIBMESH_DIM variable sets such as displacements.

There was also a minor bug fix in this commit. The passed-in size
of the dual numbers container may be different from the size of the
variables container.