Make MortarFrictionalPressureVectorAux compute at TIMESTEP_END
The test frictional-mortar-3d-status had been intermittently failing with components
 of the computed tangential vector that were far from the reference results. This appeared
to happen in parallel runs. The fact that the auxiliary kernel was not computing the vector
 at TIMESTEP_END did not guaranteed a converged solution. Since this aux kernel is there to help
with analysis, we can in principle just run it at TIMESTEP_END avoiding the performance hit and
generating always the converged aux kernel solution.

Credit to Alex Lindsay.