Reduce heap memory allocation/deallocation with frictional mortar
When profiling the code, the amount of memory allocation/deallocation
pertaining to standard vectors that hold tangent vectors stood out (1.5GB)
for a mid-size problem. It was by far the largest hit setting aside Jacobian
entry allocations. For this reason, this commit replaces the std::vector with
the semidynamic vector that Daniel created based on the MetaPhysicL library.
 This change should have no effect other than better memory management.