Enable manual build of PETSc and HDF5 on Apple Silicon (#18954, #20736)
This requires a configuration step within the downloaded HDF5. A .patch file
is added to moose/scripts/ to facilitate this, and logic is added within
configure_petsc.sh to monitor the patch status and apply the patch during
PETSc configuration.

MUMPS also requires using the environment variable FFLAGS in order to build
correctly on arm64 when running the update_and_rebuild_petsc script. This
flag is added during configure for that specific case (the conda version of
the configure line already does this).