Implement Coupleable::vectorTagDofValueHelper
In order to make this work I had to make a virtual interface at the
MooseVariableField level for vectorTagDofValue, which led to me
implementing that method for finite volume variables, and then testing
it. In testing, I discovered that it can be dangerous to add old
solution state tags to the Coupleable _fe_coupleable_vector_tags
member. The danger comes from the fact that members like _vector_tag_u
in MooseVariableData(FV) are sized at construction time and this may
happen before tags for old solution states are added. So if we attempt
to index into _vector_tag_u with the old solution state tag, this can
lead to out-of-bounds access. However, not to fear... there is no
conceptual need for having an entry in _vector_tag_u for old solution
states for which we already have dedicated data members like _dof_values_old,
_dof_values_older, _dof_values_previous_nl. So I created a _special_tags
data member in Coupleable corresponding to these data types which
allows us to avoid indexing _vector_tag_u out-of-bounds

Other notes:
- Made solution state methods in SystemBase virtual such that the
  displaced systems can call their undisplaced system counterpart
  methods
- Fixed a bug in vectorTagValue in the variable classes where we were
  checking for OLD in the OLDER branch