Change advection schemes test to have odd number of elements
This is because of our code
```c++
template <typename Scalar, typename Vector>
Scalar
rF(const Scalar & phiC, const Scalar & phiD, const Vector & gradC, const RealVectorValue & dCD)
{
  if ((phiD - phiC) == 0)
    return 1e6 * MathUtils::sign(gradC * dCD) + 0 * (gradC * dCD + phiD + phiC);

  return 2. * gradC * dCD / (phiD - phiC) - 1.;
}
```

Our test here is symmetric in the x-velocity variable about the center
of the domain so we trigger the if condition. I found that if I change
from 1e6 to 1 (which corresponds to central differencing) then nonlinear
convergence is still not good. Note that vanLeer and min_mod do not
have issues with this setup, presumably because they bound the value
of r. Anyway I'm going to make this test use an odd number of cells