Pre-request old solution state in constructor
Antonio was getting hangs (before my `libmesh_parallel_only`
assertion addition) because one process was requesting the old
solution vector while another process was not. And then when
we were updating our system vectors at timestep end, we would
be trying to scatter ghosts from one vector on one process and
a different vector on another process, which lead to the hang.
By requesting the old solution state in sync on all processes,
we avoid that issue