Iterate by insertion order
This is necessary for being in sync in multi-process runs. This
has a cost associated with it, e.g. when we are iterating over
the insertion order to begin our depth-first search we have to
perform O(log n) look-ups in the adjacency map, but hopefully
this will not show up as too much of a penalty