Revert mark mortar constraint variable as active in fe_problem (#19444)
These lines of code make assessment cases leveraging mortar constraints to
model gap heat transfer. It appears that problem variables (temperature) are
not properly intitialized when reinitializing face materials from the mortar
constraint loop. The ultimate cause of this issue is not known, but it's not
straightforward to reproduce in a MOOSE regression test.