Eliminate presence of empty secondary elements.
There will be nodes that do not project onto the primary surface. This will
cause that maps in loopOverMortarSegments that do not have elements in the mortar segment
mesh. As such, no integration points are identified, which will cause issues when reiniting
entities such as dual bases. Here we eliminate empty secondaries in AutomaticMortarGeneration::secondariesToMortarSegments.