Add contact test.
It shows that the same contact results can be obtained with a wrap around definition of
 the subdomains but with a sideset that only spans the possible contact area. This is not
true right now for a case where the sideset also wraps around since, when performing projections,
large gaps (not in contact) in many mortar segments have more weight than negative/zero gaps (in contact).
This is a corner case for a square geometry that may need special consideration for the node-based approach
used in mortar mechanical contact.