Resolved issue where fluid properties are initialized twice, reading the csv file twice, causing them to error.
TFP can now run without dependence on another fp object, a csv file with data is all that is needed for this method to work

When variable set inversion of (p,T) from (v,e) or (v,h) occurs while using csv file, it is possible there will be (p,T) values that are out of bounds (converting from a (v,e) or (v,h) grid). When this is the case, pressure and temperature are reset to their respective range min or max values depending on the situation.

ref #20101