Add comments on why not default mortar object to true.
The AugmentSparsityOnInterface object relies on a correct object
setting for _use_displace_mesh. If we override this parameter to
true in cases where there is no displaced mesh, we will be trying to
retrieve a mortar mesh that does not exist. This happens because
FeProblemBase changes the value of this parameter externally.