Publish coupling information to process owning rows
It is not good enough to add PrimaryLower coupling to the process
which owns the *lower* dofs. libMesh fills the sparsity pattern by
looping over local elements and then fills the j-column entries
of the i-rows associated with the local elements by calling the coupling
functor on each process with its local elements. So we have to publish
PrimaryLower and PrimarySecondary coupling information to the processes
that own the Primary dofs (and similarly SecondaryLower and SecondaryPrimary
information to the processes that own Secondary dofs)