Revert "Update VariableValueSamplePostprocessorTransfer to MooseCoordTransform"
This transfer relies on querying the positions/translations of all sub-apps in
the MultiApp, regardless of whether they are local to a process or not.  If a
sub-app is not local then we have no access to its FEProblemBase and
consequently we do not have access to a coordinate transformation object.
Scaling and rotation do not affect the output of MooseCoordTransform::operator()
when the argument is the origin ... but we perform coordinate collapsing
post-translation. I do not want to add some check that errors out if a MultiApp
has a coordinate system such that coordinate collapsing occurs, in which case
`position` potentially does not match MooseCoordTransform::operator()(Point(0)).

This reverts commit 60afbb887cbe96020ecb39c03daa45bd10ae561e.