Cartesian LMs are an alternative to local contact frame LMs (#21461)
Initially we thought of fully substituting the local frame formulation by
the Cartesial LMs approach. But, due to the extent of development of local LMs'
mechanical contact, it may be more straightforward to allow both approaches to
coexist while we further test the implementation for accuracy, condensation,
usage of iterative solvers.

This PR also forces the Cartesian objects to avoid interpolating geometry. In other words,
it uses a single local frame for each constraint. This prevents deviations from orthonormal
frames while interpolating geometry within an element. This may not be the source of
a lot of error in general, so we may consider also adding the interpolation option in the
future.